# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RiverKin API",
    "description": "RiverKin backend \u2014 IEEE OneAquaHealth Global Hackathon 2026 (Track 5).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/sites": {
      "get": {
        "tags": [
          "sites"
        ],
        "summary": "List Sites",
        "description": "All sites, most-urgent first. Optional ``city`` filter.",
        "operationId": "list_sites_api_v1_sites_get",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SiteOut"
                  },
                  "title": "Response List Sites Api V1 Sites Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sites/{oah_code}": {
      "get": {
        "tags": [
          "sites"
        ],
        "summary": "Get Site",
        "operationId": "get_site_api_v1_sites__oah_code__get",
        "parameters": [
          {
            "name": "oah_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Oah Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sites/{oah_code}/timeline": {
      "get": {
        "tags": [
          "sites"
        ],
        "summary": "Site Timeline",
        "description": "Checks, rain context and verifications for a site, newest first (C7).\n\nCombines real submitted observations with the site's rain/last-check context.",
        "operationId": "site_timeline_api_v1_sites__oah_code__timeline_get",
        "parameters": [
          {
            "name": "oah_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Oah Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteTimelineOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/verify/next": {
      "get": {
        "tags": [
          "verify"
        ],
        "summary": "Next Verify Items",
        "description": "Return the next verify cards for this voter.\n\nSkips items the voter already voted on and items whose observation is no\nlonger open. Gold items are included but never marked as gold in the payload.",
        "operationId": "next_verify_items_api_v1_verify_next_get",
        "parameters": [
          {
            "name": "n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 5,
              "title": "N"
            }
          },
          {
            "name": "voter_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-keeper",
              "title": "Voter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyNextOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/verify/{item_id}/vote": {
      "post": {
        "tags": [
          "verify"
        ],
        "summary": "Cast Vote",
        "description": "Record a verification vote and recompute the observation's trust + state.",
        "operationId": "cast_vote_api_v1_verify__item_id__vote_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoteIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SiteOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "waterbody": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Waterbody"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lng"
          },
          "days_unseen": {
            "type": "integer",
            "title": "Days Unseen"
          },
          "rain_48h_mm": {
            "type": "number",
            "title": "Rain 48H Mm"
          },
          "need_score": {
            "type": "number",
            "title": "Need Score"
          },
          "attention": {
            "type": "string",
            "title": "Attention"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "days_unseen",
          "rain_48h_mm",
          "need_score",
          "attention",
          "color"
        ],
        "title": "SiteOut",
        "description": "A monitored OAH site with its current attention state.\n\n``id`` is the stable OAH code (string) so the front end can route on it."
      },
      "SiteTimelineOut": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/TimelineEntry"
            },
            "type": "array",
            "title": "Entries"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "entries"
        ],
        "title": "SiteTimelineOut"
      },
      "TimelineEntry": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "at": {
            "type": "string",
            "title": "At"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "label",
          "at"
        ],
        "title": "TimelineEntry"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyCard": {
        "properties": {
          "item_id": {
            "type": "integer",
            "title": "Item Id"
          },
          "observation_id": {
            "type": "integer",
            "title": "Observation Id"
          },
          "site_name": {
            "type": "string",
            "title": "Site Name"
          },
          "field_code": {
            "type": "string",
            "title": "Field Code"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "ai_box": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Box"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "observation_id",
          "site_name",
          "field_code",
          "question"
        ],
        "title": "VerifyCard",
        "description": "One verify card for C5. Submitter, tally and gold status are hidden."
      },
      "VerifyNextOut": {
        "properties": {
          "cards": {
            "items": {
              "$ref": "#/components/schemas/VerifyCard"
            },
            "type": "array",
            "title": "Cards"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "cards"
        ],
        "title": "VerifyNextOut"
      },
      "VoteIn": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "ms_taken": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ms Taken"
          },
          "voter_kind": {
            "type": "string",
            "title": "Voter Kind",
            "default": "keeper"
          },
          "voter_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voter Id"
          }
        },
        "type": "object",
        "required": [
          "answer"
        ],
        "title": "VoteIn"
      },
      "VoteResult": {
        "properties": {
          "recorded": {
            "type": "boolean",
            "title": "Recorded"
          },
          "observation_id": {
            "type": "integer",
            "title": "Observation Id"
          },
          "observation_status": {
            "type": "string",
            "title": "Observation Status"
          },
          "trust": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trust"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "recorded",
          "observation_id",
          "observation_status",
          "trust"
        ],
        "title": "VoteResult"
      }
    }
  }
}